Extract isolation core into dedicated module
